+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Changes
 
 * Make `YARDSorbet::SigToYARD::IS_LEGACY_RUBY_VERSION` private
+* Minimum `sorbet` version increased to `0.5.6289`
+* Minimum `yard` version increased to `0.9.21`
 
 ## 0.4.1 (2021-04-28)